Update files affected by changes
